+++ src/app/common/helpers/base-test.helpers.spec.ts
@@ -158,6 +158,7 @@ export function registerMatSvgIcons(): void {
         "hour-glass",
         "compact",
         "access-token",
+        "tree-structure",
     ];
     icons.forEach((icon: string) => {
         matIconRegistry.addSvgIcon(icon, domSanitizer.bypassSecurityTrustResourceUrl("/fake-path/fake-icon.svg"));
