+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/transform-options/dataset-settings-transform-options-tab.component.html
@@ -1,4 +1,4 @@
-<div class="custom-container">
+<div class="custom-container pb-80">
     <h2>Transform settings</h2>
     <mat-divider class="mt-1 mb-2" />
