+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/transform-options/dataset-settings-transform-options-tab.component.scss
@@ -1,0 +1,3 @@
+.pb-80 {
+    padding-bottom: 80px;
+}
